requirements.txt
alembic==1.12.0
apispec==6.3.0
async-timeout==4.0.3
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==2.2.2
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
flask-smorest==0.42.1
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
passlib==1.7.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
rq==1.15.1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
webargs==8.3.0
Werkzeug==2.3.7
